feed ui done

+++ app/src/main/java/in/bitotsav/bitotsav_20/feed/data/FeedRepository.kt
@@ -20,5 +20,5 @@ class FeedRepository(private val feedDao: FeedDao) {
     fun getAll() = feedDao.getAll()
 
     @WorkerThread
-    suspend fun getFeed(id: Long) = feedDao.get(id)
+    suspend fun getFeed(id: Int) = feedDao.get(id)
 }
